+++ ExcelWakaTimeAddin/Forms/SettingsForm.cs
@@ -1,16 +1,20 @@
 ﻿using System;
+using System.Text.RegularExpressions;
 using System.Windows.Forms;
+using WakaTime.Shared.ExtensionUtils;
 
 namespace WakaTime.Forms
 {
     public partial class SettingsForm : Form
     {
-        private readonly Shared.ExtensionUtils.WakaTime _wakaTime;
-        internal event EventHandler ConfigSaved;
+        private readonly ConfigFile _configFile;
+        private readonly ILogger _logger;
 
-        public SettingsForm(ref Shared.ExtensionUtils.WakaTime wakaTime)
+        public SettingsForm(ConfigFile configFile, ILogger logger)
         {
-            _wakaTime = wakaTime;
+            _configFile = configFile;
+            _logger = logger;
+
             InitializeComponent();
         }
 
@@ -18,48 +22,43 @@ namespace WakaTime.Forms
         {
             try
             {
-                txtAPIKey.Text = _wakaTime.Config.ApiKey;
-                txtProxy.Text = _wakaTime.Config.Proxy;
-                chkDebugMode.Checked = _wakaTime.Config.Debug;
+                txtAPIKey.Text = _configFile.GetSetting("api_key");
+                txtProxy.Text = _configFile.GetSetting("proxy");
+                chkDebugMode.Checked = _configFile.GetSettingAsBoolean("debug");
             }
             catch (Exception ex)
             {
-                _wakaTime.Logger.Error("Error when loading form SettingsForm:", ex);
+                _logger.Error("Error when loading form SettingsForm:", ex);
+
                 MessageBox.Show(ex.Message);
             }
         }
 
-        private void btnOk_Click(object sender, EventArgs e)
+        private void BtnOk_Click(object sender, EventArgs e)
         {
             try
             {
-                var parse = Guid.TryParse(txtAPIKey.Text.Trim(), out var apiKey);         
-                                     
-                if (parse)
+                var matched = Regex.IsMatch(txtAPIKey.Text.Trim(), "(?im)^(waka_)?[0-9A-F]{8}[-]?(?:[0-9A-F]{4}[-]?){3}[0-9A-F]{12}$");
+
+                if (matched)
                 {
-                    _wakaTime.Config.ApiKey = apiKey.ToString();
-                    _wakaTime.Config.Proxy = txtProxy.Text.Trim();
-                    _wakaTime.Config.Debug = chkDebugMode.Checked;
-                    _wakaTime.Config.Save();
-                    OnConfigSaved();
+                    _configFile.SaveSetting("settings", "api_key", txtAPIKey.Text.Trim());
+                    _configFile.SaveSetting("settings", "proxy", txtProxy.Text.Trim());
+                    _configFile.SaveSetting("settings", "debug", chkDebugMode.Checked.ToString().ToLower());
                 }
                 else
                 {
                     MessageBox.Show(@"Please enter valid Api Key.");
+
                     DialogResult = DialogResult.None; // do not close dialog box
                 }
             }
             catch (Exception ex)
             {
-                _wakaTime.Logger.Error("Error when saving data from SettingsForm:", ex);
+                _logger.Error($"Error saving data from SettingsForm: {ex}");
+
                 MessageBox.Show(ex.Message);
             }
-        }
-
-        protected virtual void OnConfigSaved()
-        {
-            var handler = ConfigSaved;
-            handler?.Invoke(this, EventArgs.Empty);
         }
     }
 }